refactor: relocate web token setup
